agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-community>=0.3.23",
    "langchain>=0.3.25",
    "streamlit>=1.45.0",
    "openai>=1.78.0",
    "langchain-openai>=0.3.16",
    "google-generativeai>=0.8.5",
    "fastapi[standard]>=0.115.12",
]
